refactor(interbank): split gRPC connections per service + real permission provider
Two minor follow-ups to align with the banking-service layout.

1. connections.go split into one file per external service
   (user_service.go, trading_service.go) so the layout matches the
   existing pattern in banking-service/internal/client.

2. Replace NoopPermissionProvider with a real gRPC-backed one
   (internal/permission/grpc_provider.go, copied from banking) that
   resolves permissions via the PermissionService exposed by
   user-service. Shares the user-service gRPC connection — no new
   network endpoint. Keeps the door open for handler-level
   auth.RequirePermission(...) checks down the line.
